refactor(issue): move search issue to separate file
